requirements.txt
pytorch
numpy